Merged PR 1748932: Making the app look awesome for Build.
The app is now looking like what Conroy envisioned. Not complete but pretty good.
Updated to likely Win2d shipping candidate.
Refactor of CompositionPlayer to eliminate VisualPlayer. The goal of this was to allow other CompositionObjects to animate in synchrony with CompositionPlayer. Currently used to show progress in the scrubber of the app.
Improvements to the canonicalizer so we have more efficient generated code.
Started on the C++ code generator.
Implemented animated splash screen (not really a splash screen, but it looks like it is).
Plus a bunch more improvements.